Renamed Notes_Controller.py to Operations_Controller.py Added Classes to Operations_Controller Added comments to Operations_Controller Added minor changes to Operations_Controller
Added Add_Note.py to Operations folder, for testing pursposes (has no functionality outside of printing)
