+++ modules/Notes Module/Operations/Add_Note.py
@@ -1,0 +1,2 @@
+def __init__():
+    return "Add Note Operation"
